+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.0.3]
+
+1. Added support for android 12
+2. Added BLUETOOTH_CONNECT permission for android 12 onwards
+3. Added isPermissionBluetoothGranted function to detect if permission is enabled, works only on android 12 and up
+4. Changed Kotlin from 1.3.50 to 1.6.10
+
 ## [1.0.2]
 
 Added a method to disconnect the printer